+++ changelog.md
@@ -2,6 +2,7 @@ v4.6.0
 * Delete Routine for exadelete Block
 * new Grade Reports for Final Grades
 * other new Reports (Abgangszeugnis, Abschlusszeugnis, Zertifikat Profilfach) 
+* New Grade Overview Report (Notenspiegel)
 
 v4.5.1
 * bugfixes